requirements.txt
groq>=0.9.0
python-dotenv>=1.0
streamlit>=1.30
pytest>=7.0
